requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django ~> 2.0.4

# Cross request requests
django-cors-headers


django-filter==1.1.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.3
Markdown==2.6.11
pytz==2018.4
static3==0.7.0
web.py==0.40.dev1

# Despliegue en heroku
django-heroku
gunicorn
whitenoise

# Postgress
psycopg2-binary

#Mysql database connection
cymysql
django-cymysql
